requirements.txt
fastapi
uvicorn
GitPython
pydriller
semgrep
mistralai
python-dotenv
pydantic